Bug 830265 make help link point to suse docs
Added a dashboard with links to the installed SUSE manuals and the
upstream documentation. Replaced the old "Help" link with a link to
the help page. The links are broken untill ibs sr#28723 is accepted

Mark the susedoc dashboard as a module

Added dashboard declaration

Removed Deployment guide from the help page